app/pyproject.toml
[tool.poetry]
name = "apicvevulnerabilities"
version = "0.1.0"
description = "API vulnerabilities CVE"
authors = ["kurotom"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
whitenoise = "^6.3.0"
gunicorn = "^20.1.0"
django-allauth = "^0.52.0"
dj-rest-auth = "^3.0.0"
drf-yasg = "^1.21.5"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
